# docs/yaml/openapi.yaml
openapi: 3.0.0
info:
   version: 1.0.0
   title: Back End Demonstration API
   description: A simple blog API.
   contact:
      name: Ben Tung
      email: 227019974@stu.vtc.edu.hk
   license:
      name: Apache 2.0
      url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://lab6blog.bentung.repl.co'
    description: Development Server
tags:
  - name: Users
    description: API endpoints for user management.
  - name: Articles
    description: Access and perform CRUD operations on blog entries.
paths:
  /articles:
    get:
      tags:
        - Articles
      description: Get all articles
      responses:
        '200':
          description: Request success
    post:
      tags:
        - Articles
      description: Create article
      requestBody:
        description: A new article data
        required: true
        content:
          application/json:
            schema:
              $ref: ./article.json#/definitions/article
        responses:
          '200':
            description: Create article successfully
          '400':
            description: Create article unsuccessfully
  /articles/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the article
    get:
      tags:
        - Articles
      description: Get article by id.
      responses:
        '200':
          description: View of an article record
        '404':
          description: Article was not found in the DB
    put:
      tags:
        - Articles
      description: Update article by id.
      requestBody:
        description: A new article data
        required: true
        content:
          application/json:
            schema:
              $ref: ./article.json#/definitions/article
        responses:
          '200':
            description: Update designed article successfully
          '400':
            description: Update designed article unsuccessfully
    delete:
      tags:
        - Articles
      description: Delete article by ID.
      